Set `Collection#creation_date` when doing `#apply_vdc_metadata`
Creation date on collections is to be system assigned and stored as a `Date` (no
time or zone information). This is implemented here per #3.

`Collection#apply_vdc_metadata` metadata is called during `#create` controller
actions. Inlining the `#assign_creation_date` process there is the most
expedient way to get the date set.

We still need to:

  - [ ] test controllers
  - [ ] remove `#creation_date` from collection forms
  - [ ] test display for ISO 8601 format