Send mail notifications to admin users when creating large collections
Wires the existing collection size mailer into the collections controller(s) to
ensure notifications get sent to admins when a collection created or updated to
the one of the "large settings". "Large" is defined by
`CollectionSizeNotification::THRESHOLD`, which needs to be maintained when/if the
vocabulary ids change, but provides an interface (`#include?`) for a more
resiliant approach if change becomes an issue.

`CollectionSizeNotification` exists mainly to encapsulate the branch
condition/decision process for whether to send a notification for a given
collection.

Connected to #16.