Setup separate Solr core for test environment
The Solr core being used for `test` was the same as the `development` core. This
led routine test cleanup to clear the development index.

Using a separate core requires special setup (documented in
`DOCKER_DEVELOPMENT`), but keeps the test environment properly partitioned. The
`SOLR_URL` environment variable is used as a fallback, to simplify CI setup.