Configure for use with `mysql2` and dockerize MySQL 5.7
Unfortunately, several existing migrations cause exceptions when run with the
`mysql2` adapter. One of these predates Notch8's work; the changes here reflect
what's in place on the pre-production server. The other should be isolated to
Notch8 development environments. Since we are moving those to MySQL, it's
application should be transparent.

Developers will need to redo the steps in `DOCKER_DEVELOPMENT.md` after pulling
this work.