Handle `nil` discipline cases
The codebase counts on discipline existing, and previously had a default value:
`[]`. However the `pre-prod` codebase does not enforce this at the database
level, leading to all manner of errors! This handles the nil case instead of
taking it on faith.

This is admittedly a slightly hacky fix, but seriously improves on what's
in (pre-) production now, and at minimal cost.