Display full collection titles
The `Vdc::CollectionForm` has some code to make the `:title` field
singular. This runs into conflict with Hyrax, which wants to use
`@from.title.first` to display the title. Hyrax shouldn't do this, but there's
no easy fix. We can't return an `Array` or `ActiveTriples::Relation`, because
this will be treated as the (singular) value of the field. We can't return a
`String` or we'll end up displaying only the first character.

Instead, we wrap the title in a delegator which does both things at once!

Fixes #37.