Add Person Indexing and search
Adds `Vdc::Person` models to the scope of the `CatalogSearchBuilder`. To do
this, we add a custom search builder class, inheriting from the Hyrax
default. We inline VDC's existing customizations there to avoid load order
issues, developer confusion, and just plain complexity.

A very simple view partial for the search results is added, as well as a helper
pointing it to the user profile.

Some index fields are also added, to put some basic information in the search
results and power simple search against people.

Connected to #10.